chore: setup linting and fix resulting issues
