+++ ark-taco-shop/.prettierignore
@@ -1,0 +1,3 @@
+.coverage
+dist
+docs
