+++ ark-taco-shop/.prettierrc.json
@@ -1,0 +1,7 @@
+{
+    "printWidth": 120,
+    "singleQuote": false,
+    "tabWidth": 4,
+    "trailingComma": "all",
+    "useTabs": false
+}
