+++ ark-taco-shop/package.json
@@ -1,62 +1,48 @@
 {
-  "name": "@mlh/ark-taco-shop",
-  "description": "Provides a Taco Shop UI for Ark Core",
-  "version": "0.1.0",
-  "private": true,
-  "contributors": [
-    "Halisson Bruno <halissonvit@gmail.com>"
-  ],
-  "license": "MIT",
-  "main": "dist/index",
-  "types": "dist/index",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "start": "tsc && node ./bin/www",
-    "build": "yarn compile",
-    "compile": "./node_modules/typescript/bin/tsc",
-    "dev": "cross-env DEBUG=ark-taco-shop:* nodemon bin/www",
-    "build:docs": "../../node_modules/.bin/jsdoc -c jsdoc.json",
-    "test": "ARK_ENV=test jest --runInBand --detectOpenHandles",
-    "test:coverage": "ARK_ENV=test jest --coverage --runInBand --detectOpenHandles",
-    "test:debug": "ARK_ENV=test node --inspect-brk ../../node_modules/.bin/jest --runInBand --watch",
-    "test:watch": "ARK_ENV=test jest --runInBand --watch",
-    "test:watch:all": "ARK_ENV=test jest --runInBand --watchAll",
-    "lint": "eslint ./ --fix",
-    "depcheck": "depcheck ./"
-  },
-  "devDependencies": {
-    "@arkecosystem/core-interfaces": "^2.2.1",
-    "@sindresorhus/tsconfig": "^0.2.1",
-    "@types/express": "^4.16.1",
-    "@types/http-proxy-middleware": "^0.19.2",
-    "@types/node": "^11.10.5",
-    "cross-env": "^5.2.0",
-    "depcheck": "^0.6.9",
-    "eslint": "^4.19.1",
-    "eslint-config-standard": "^11.0.0",
-    "eslint-plugin-import": "^2.12.0",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-standard": "^3.1.0",
-    "jest": "^23.1.0",
-    "jest-extended": "^0.7.2",
-    "jsdoc": "^3.5.5",
-    "regenerator-runtime": "^0.11.1",
-    "standard": "^11.0.1",
-    "typescript": "^3.3.3333",
-    "winston": "^3.1.0"
-  },
-  "dependencies": {
-    "@arkecosystem/client": "^0.1.20",
-    "@arkecosystem/crypto": "^2.1.1",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
-    "express": "~4.16.0",
-    "http-errors": "~1.6.2",
-    "http-proxy-middleware": "^0.19.0",
-    "morgan": "~1.9.0",
-    "nodemon": "^1.18.10"
-  }
+    "name": "@mlh/ark-taco-shop",
+    "description": "Provides a Taco Shop UI for Ark Core",
+    "version": "0.1.0",
+    "private": true,
+    "contributors": [
+        "Halisson Bruno <halissonvit@gmail.com>"
+    ],
+    "license": "MIT",
+    "main": "dist/index",
+    "types": "dist/index",
+    "files": [
+        "dist"
+    ],
+    "scripts": {
+        "start": "yarn compile && node ./bin/www",
+        "build": "yarn compile",
+        "compile": "./node_modules/typescript/bin/tsc",
+        "dev": "cross-env DEBUG=ark-taco-shop:* nodemon bin/www",
+        "format": "./node_modules/tslint/bin/tslint -c ./tslint.json 'src/**/*.ts' --fix && prettier --write \"./*.{ts,js,json,md}\" \"./src/**/*.{ts,js,json,md}\""
+    },
+    "dependencies": {
+        "@arkecosystem/client": "^0.1.20",
+        "@arkecosystem/crypto": "^2.1.1",
+        "cookie-parser": "~1.4.3",
+        "debug": "~2.6.9",
+        "express": "~4.16.0",
+        "http-errors": "~1.6.2",
+        "http-proxy-middleware": "^0.19.0",
+        "morgan": "~1.9.0",
+        "nodemon": "^1.18.10"
+    },
+    "devDependencies": {
+        "@arkecosystem/core-interfaces": "^2.2.1",
+        "@sindresorhus/tsconfig": "^0.2.1",
+        "@types/express": "^4.16.1",
+        "@types/http-proxy-middleware": "^0.19.2",
+        "@types/node": "^11.10.5",
+        "cross-env": "^5.2.0",
+        "jest": "^23.1.0",
+        "jest-extended": "^0.7.2",
+        "prettier": "^1.16.4",
+        "tslint": "^5.14.0",
+        "tslint-config-prettier": "^1.18.0",
+        "typescript": "^3.3.3333",
+        "winston": "^3.1.0"
+    }
 }
