+++ ark-taco-shop/tsconfig.json
@@ -1,18 +1,18 @@
 {
-  "extends": "@sindresorhus/tsconfig",
-  "compilerOptions": {
-    "lib": ["es2018"],
-    "typeRoots": ["node_modules/@types"],
-    "target": "es2018",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "outDir": "dist",
-    "sourceMap": true,
-    "strict": false,
-    "noUnusedLocals": false,
-    "noUnusedParameters": false,
-    "resolveJsonModule": true
-  },
-  "include": ["lib/**/**.ts"],
-  "exclude": ["node_modules", "**/*.spec.ts", "**/*.test.ts"]
+    "extends": "@sindresorhus/tsconfig",
+    "compilerOptions": {
+        "lib": ["es2018"],
+        "typeRoots": ["node_modules/@types"],
+        "target": "es2018",
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "sourceMap": true,
+        "strict": false,
+        "noUnusedLocals": false,
+        "noUnusedParameters": false,
+        "resolveJsonModule": true,
+        "outDir": "dist"
+    },
+    "include": ["src/**/**.ts"],
+    "exclude": ["node_modules", "**/*.spec.ts", "**/*.test.ts"]
 }
