+++ ark-taco-shop/tslint.json
@@ -1,0 +1,11 @@
+{
+    "extends": ["tslint:recommended", "tslint-config-prettier"],
+    "rules": {
+        "interface-name": false,
+        "no-console": false,
+        "no-default-export": true,
+        "no-shadowed-variable": false,
+        "object-literal-sort-keys": false,
+        "radix": false
+    }
+}
